Extract App jsx as a separate file
