+++ src/index.jsx
@@ -1,9 +1,5 @@
 import React from 'react';
 import { render } from 'react-dom';
-import { Provider } from 'react-redux';
-import { BrowserRouter, Route, Switch } from 'react-router-dom';
-import { ThemeProvider } from 'styled-components';
-import querystring from 'querystring';
 import { library } from '@fortawesome/fontawesome-svg-core';
 import {
   faAngleUp,
@@ -13,8 +9,6 @@ import {
   faUser,
 } from '@fortawesome/free-solid-svg-icons';
 import ReactGA from 'react-ga';
-import { Helmet } from 'react-helmet';
-
 import '@gen3/ui-component/dist/css/base.less';
 import {
   fetchDictionary,
@@ -24,42 +18,10 @@ import {
   fetchUserAccess,
   fetchUserAuthMapping,
 } from './actions';
-import ReduxLogin, { fetchLogin } from './Login/ReduxLogin';
-import ProtectedContent from './Login/ProtectedContent';
-import SubmissionPage from './Submission/page';
-import DocumentPage from './Document/page';
-import CoreMetadataPage from './CoreMetadata/page';
-import { fetchCoreMetadata } from './CoreMetadata/reduxer';
-import Indexing from './Indexing/Indexing';
-import IndexPage from './Index/page';
-import DataDictionary from './DataDictionary/.';
-import PrivacyPolicy from './PrivacyPolicy/PrivacyPolicy';
-import ProjectSubmission from './Submission/ReduxProjectSubmission';
-import ReduxMapFiles from './Submission/ReduxMapFiles';
-import ReduxMapDataModel from './Submission/ReduxMapDataModel';
-import UserProfile, { fetchAccess } from './UserProfile/ReduxUserProfile';
-import UserAgreementCert from './UserAgreement/ReduxCertPopup';
-import GraphQLQuery from './GraphQLEditor/ReduxGqlEditor';
-import theme from './theme';
 import getReduxStore from './reduxStore';
-import { ReduxNavBar, ReduxTopBar, ReduxFooter } from './Layout/reduxer';
-import ReduxQueryNode, { submitSearchForm } from './QueryNode/ReduxQueryNode';
-import {
-  basename,
-  dev,
-  gaDebug,
-  workspaceUrl,
-  workspaceErrorUrl,
-  enableResourceBrowser,
-} from './localconf';
-import { gaTracking, components } from './params';
-import GA, { RouteTracker } from './components/GoogleAnalytics';
-import GuppyDataExplorer from './GuppyDataExplorer/.';
-import isEnabled from './helpers/featureFlags';
+import { gaTracking } from './params';
+import App from './App';
 import sessionMonitor from './SessionMonitor';
-import Workspace from './Workspace';
-import ResourceBrowser from './ResourceBrowser';
-import ErrorWorkspacePlaceholder from './Workspace/ErrorWorkspacePlaceholder';
 import './index.less';
 
 // monitor user's session
@@ -87,213 +49,7 @@ async function init() {
     store.dispatch(fetchProjects()),
   ]);
 
-  render(
-    <Provider store={store}>
-      <ThemeProvider theme={theme}>
-        <BrowserRouter basename={basename}>
-          {GA.init(gaTracking, dev, gaDebug) && <RouteTracker />}
-          {isEnabled('noIndex') && (
-            <Helmet>
-              <meta name='robots' content='noindex,nofollow' />
-            </Helmet>
-          )}
-          <ReduxTopBar />
-          <ReduxNavBar />
-          <Switch>
-            <Route
-              path='/login'
-              component={(props) => (
-                <ProtectedContent
-                  isPublic
-                  filter={() => store.dispatch(fetchLogin())}
-                  component={ReduxLogin}
-                  {...props}
-                />
-              )}
-            />
-            <Route
-              exact
-              path='/'
-              component={(props) => (
-                <ProtectedContent component={IndexPage} {...props} />
-              )}
-            />
-            <Route
-              exact
-              path='/submission'
-              component={(props) => (
-                <ProtectedContent
-                  isAdminOnly
-                  component={SubmissionPage}
-                  {...props}
-                />
-              )}
-            />
-            <Route
-              exact
-              path='/submission/files'
-              component={(props) => (
-                <ProtectedContent
-                  isAdminOnly
-                  component={ReduxMapFiles}
-                  {...props}
-                />
-              )}
-            />
-            <Route
-              exact
-              path='/submission/map'
-              component={(props) => (
-                <ProtectedContent
-                  isAdminOnly
-                  component={ReduxMapDataModel}
-                  {...props}
-                />
-              )}
-            />
-            <Route
-              exact
-              path='/document'
-              component={(props) => (
-                <ProtectedContent component={DocumentPage} {...props} />
-              )}
-            />
-            <Route
-              path='/query'
-              component={(props) => (
-                <ProtectedContent component={GraphQLQuery} {...props} />
-              )}
-            />
-            <Route
-              path='/identity'
-              component={(props) => (
-                <ProtectedContent
-                  filter={() => store.dispatch(fetchAccess())}
-                  component={UserProfile}
-                  {...props}
-                />
-              )}
-            />
-            <Route
-              path='/indexing'
-              component={(props) => (
-                <ProtectedContent component={Indexing} {...props} />
-              )}
-            />
-            <Route
-              path='/quiz'
-              component={(props) => (
-                <ProtectedContent component={UserAgreementCert} {...props} />
-              )}
-            />
-            <Route
-              path='/dd/:node'
-              component={(props) => (
-                <ProtectedContent component={DataDictionary} {...props} />
-              )}
-            />
-            <Route
-              path='/dd'
-              component={(props) => (
-                <ProtectedContent component={DataDictionary} {...props} />
-              )}
-            />
-            <Route
-              exact
-              path='/files/*'
-              component={(props) => (
-                <ProtectedContent
-                  filter={() =>
-                    store.dispatch(fetchCoreMetadata(props.match.params[0]))
-                  }
-                  component={CoreMetadataPage}
-                  {...props}
-                />
-              )}
-            />
-            <Route
-              path='/files'
-              component={(props) => (
-                <ProtectedContent component={GuppyDataExplorer} {...props} />
-              )}
-            />
-            <Route
-              path='/workspace'
-              component={(props) => (
-                <ProtectedContent component={Workspace} {...props} />
-              )}
-            />
-            <Route path={workspaceUrl} component={ErrorWorkspacePlaceholder} />
-            <Route
-              path={workspaceErrorUrl}
-              component={ErrorWorkspacePlaceholder}
-            />
-            <Route
-              path='/:project/search'
-              component={(props) => {
-                const queryFilter = () => {
-                  const location = props.location;
-                  const queryParams = querystring.parse(
-                    location.search ? location.search.replace(/^\?+/, '') : ''
-                  );
-                  if (Object.keys(queryParams).length > 0) {
-                    // Linking directly to a search result,
-                    // so kick-off search here (rather than on button click)
-                    return store.dispatch(
-                      submitSearchForm({
-                        project: props.match.params.project,
-                        ...queryParams,
-                      })
-                    );
-                  }
-                  return Promise.resolve('ok');
-                };
-                return (
-                  <ProtectedContent
-                    filter={queryFilter}
-                    component={ReduxQueryNode}
-                    {...props}
-                  />
-                );
-              }}
-            />
-            {isEnabled('explorer') && (
-              <Route
-                path='/explorer'
-                component={(props) => (
-                  <ProtectedContent component={GuppyDataExplorer} {...props} />
-                )}
-              />
-            )}
-            {components.privacyPolicy &&
-              (!!components.privacyPolicy.file ||
-                !!components.privacyPolicy.routeHref) && (
-                <Route path='/privacy-policy' component={PrivacyPolicy} />
-              )}
-            {enableResourceBrowser && (
-              <Route
-                path='/resource-browser'
-                component={(props) => (
-                  <ProtectedContent component={ResourceBrowser} {...props} />
-                )}
-              />
-            )}
-            <Route
-              path='/:project'
-              component={(props) => (
-                <ProtectedContent component={ProjectSubmission} {...props} />
-              )}
-            />
-          </Switch>
-          <ReduxFooter
-            logos={components.footerLogos}
-            privacyPolicy={components.privacyPolicy}
-          />
-        </BrowserRouter>
-      </ThemeProvider>
-    </Provider>,
-    document.getElementById('root')
-  );
+  render(<App store={store} />, document.getElementById('root'));
 }
 
 init();
